+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 
-## [«unknown»](https://github.com/NASA-PDS/validate/tree/«unknown») (2023-04-19)
+## [«unknown»](https://github.com/NASA-PDS/validate/tree/«unknown») (2023-05-03)
 
 [Full Changelog](https://github.com/NASA-PDS/validate/compare/v3.2.0...«unknown»)
+
+**Requirements:**
+
+- As a user, I want to be able to use both online and local schema/schematron files. [\#599](https://github.com/NASA-PDS/validate/issues/599)
 
 **Improvements:**